refactor(vdb): organize imports with barrel files (1)
