+++ libs/vdb/src/lib/tree/index.ts
@@ -1,5 +1,7 @@
 export * from './internal-node';
 export * from './leaf-node';
+export * from './node';
 export * from './root-node';
 export * from './tree';
 export * from './value-accessor';
+export * from './voxel';
